Make kinto_remote_settings an installable package
Update Docker setup accordingly to install package in container rather
than adding the directory to the PYTHONPATH
