+++ docker-compose.yml
@@ -64,6 +64,6 @@ services:
     volumes:
       - $PWD/tests:/app
       - $PWD/mail:/app/mail/
-      - $PWD/kinto_remote_settings:/app/kinto_remote_settings
+      - $PWD/kinto-remote-settings:/app/kinto-remote-settings
       - $PWD/setup.cfg:/app/setup.cfg
     command: start
