+++ kinto-remote-settings/README.rst
