+++ kinto-remote-settings/src/kinto_remote_settings/changes/views.py
