+++ kinto-remote-settings/src/kinto_remote_settings/signer/generate_keypair.py
