+++ kinto-remote-settings/src/kinto_remote_settings/signer/utils.py
