+++ kinto-remote-settings/tests/changes/__init__.py
