+++ kinto-remote-settings/tests/changes/config.ini
