+++ kinto-remote-settings/tests/changes/test_changes.py
