+++ kinto-remote-settings/tests/signer/config/autograph.yaml
