+++ kinto-remote-settings/tests/signer/config/bob.ecdsa.private.pem
