+++ kinto-remote-settings/tests/signer/config/ecdsa.public.pem
