+++ kinto-remote-settings/tests/signer/config/signer.ini
