+++ kinto-remote-settings/tests/signer/functional.py
@@ -6,7 +6,6 @@ from urllib.parse import urljoin
 import requests
 from kinto_http import Client, KintoException
 from kinto_http.patch_type import JSONPatch
-
 from kinto_remote_settings.signer.backends import local_ecdsa
 from kinto_remote_settings.signer.serializer import canonical_json
