+++ kinto-remote-settings/tests/signer/support.py
