+++ kinto-remote-settings/tests/signer/test_generate_keypair.py
