+++ kinto-remote-settings/tests/signer/test_serializer.py
