+++ kinto-remote-settings/tests/signer/test_signer_attachments.py
