+++ setup.cfg
@@ -15,7 +15,7 @@ exclude = .git,__pycache__,node_modules
 [isort]
 profile=black
 lines_after_imports=2
-src_paths=kinto_remote_settings,tests
+src_paths=kinto-remote-settings,tests
 
 [tool:pytest]
 # https://docs.pytest.org/en/6.2.x/usage.html
